refactor index JS. fix dependancy issues
