+++ lambda/custom/handlers/quizHandler.js
@@ -1,4 +1,5 @@
 const API = require('../utils/apiUtils')
+const Alexa = require("ask-sdk-core");
 
 const QuizIntentHandler = {
   canHandle({ requestEnvelope }) {
