+++ lambda/custom/index.js
@@ -1,39 +1,5 @@
 const Alexa = require("ask-sdk-core");
-const PERMISSIONS = ['alexa::alerts:reminders:skill:readwrite'];
-const utils = require('./utils/Utils')
 const handlers = require('./handlers/index')
-const quizTime = '15:00' // hardcoded for now
-
-const LaunchRequestHandler = {
-  canHandle(handlerInput) {
-    return (
-      Alexa.getRequestType(handlerInput.requestEnvelope) === "LaunchRequest"
-      );
-    },
-    async handle(handlerInput) {
-      const client = handlerInput.serviceClientFactory.getReminderManagementServiceClient();
-      const { responseBuilder } = handlerInput
-      const user_id = requestEnvelope.session.user.userId
-      const requestEnvelope = handlerInput.requestEnvelope;
-      const permissions = requestEnvelope.context.System.user.permissions
-      if (!permissions) {
-        // if no permissions, nag the user to grant them
-        return responseBuilder
-          .speak('I need permission to create reminders. Take a look at your alexa app to grant them.')
-          .withAskForPermissionsConsentCard(PERMISSIONS)
-          .getResponse()
-        }
-      // check reminders...
-      const upToDate = await utils.checkIfRemindersAreUpToDate(user_id, quizTime, client)
-      const speakOutput = upToDate ? 
-      "what would you like me to do"
-      : "It looks like there's been a change to your medication schedule. Please say update reminders to alter your reminders accordingly.";
-      return responseBuilder
-        .speak(speakOutput)
-        .reprompt(speakOutput)
-        .getResponse();
-  }
-};
 
 const HelpIntentHandler = {
   canHandle(handlerInput) {
@@ -105,10 +71,7 @@ const ErrorHandler = {
 
 exports.handler = Alexa.SkillBuilders.custom()
   .addRequestHandlers(
-    LaunchRequestHandler,
     ...handlers,
-    PairDeviceIntentHandler,
-    medsTakenHandler,
     HelpIntentHandler,
     CancelAndStopIntentHandler,
     SessionEndedRequestHandler,
